TW-934: [bug] Redux persist. + Redux DevTools
