+++ .eslintrc
@@ -43,6 +43,7 @@
         "newlines-between": "always"
       }
     ],
-    "react-hooks/exhaustive-deps": "off"
+    "react-hooks/rules-of-hooks": "warn",
+    "react-hooks/exhaustive-deps": "warn"
   }
 }
