+++ package.json
@@ -40,6 +40,7 @@
     "@ledgerhq/hw-transport-webhid": "6.27.8",
     "@peculiar/webcrypto": "1.1.6",
     "@popperjs/core": "2.4.4",
+    "@redux-devtools/remote": "^0.8.0",
     "@reduxjs/toolkit": "^1.8.5",
     "@rnw-community/shared": "^0.48.0",
     "@serh11p/jest-webextension-mock": "4.0.0",
