+++ src/app/store/buy-with-credit-card/state.ts
@@ -27,7 +27,3 @@ export const buyWithCreditCardInitialState: BuyWithCreditCardState = {
   },
   pairLimits: {}
 };
-
-export interface BuyWithCreditCardRootState {
-  buyWithCreditCard: BuyWithCreditCardState;
-}
