+++ src/app/store/root-state.mock.ts
@@ -6,9 +6,9 @@ import { mockBalancesState } from './balances/state.mock';
 import { mockBuyWithCreditCardState } from './buy-with-credit-card/state.mock';
 import { mockCurrencyState } from './currency/state.mock';
 import { mockDAppsState } from './d-apps/state.mock';
-import { RootState } from './index';
 import { mockNewsletterState } from './newsletter/newsletter-state.mock';
 import { mockPartnersPromotionState } from './partners-promotion/state.mock';
+import type { RootState } from './root-state.type';
 import { mockSettingsState } from './settings/state.mock';
 import { mockSwapState } from './swap/state.mock';
 import { mockTokensMetadataState } from './tokens-metadata/state.mock';
