+++ src/lib/store/entity.utils.ts
@@ -1,4 +1,8 @@
-import { LoadableEntityState } from '../interfaces/loadable-entity-state.interface';
+export interface LoadableEntityState<T> {
+  data: T;
+  error?: string;
+  isLoading: boolean;
+}
 
 export const createEntity = <T>(data: T, isLoading = false, error?: string): LoadableEntityState<T> => ({
   data,
