+++ src/lib/store/index.ts
@@ -1,8 +1,4 @@
-export type { LoadableEntityState } from './interfaces/loadable-entity-state.interface';
+export { createActions } from './action.utils';
 
-export type { GetStateType } from './types/get-state.type';
-
-export { createActions } from './utils/action.utils';
-export { createEntity } from './utils/entity.utils';
-export { rootStateReducer } from './utils/reducer.utils';
-export { createStore } from './utils/store.utils';
+export type { LoadableEntityState } from './entity.utils';
+export { createEntity } from './entity.utils';
