--- src/lib/store/interfaces/loadable-entity-state.interface.ts
@@ -1,5 +1,0 @@
-export interface LoadableEntityState<T> {
-  data: T;
-  error?: string;
-  isLoading: boolean;
-}
