--- src/lib/store/types/get-state.type.ts
@@ -1,3 +1,0 @@
-import { Reducer } from 'redux';
-
-export type GetStateType<R> = R extends Reducer<infer S> ? S : never;
